requirements.txt
# we need this to read/parse YAML files.
pyyaml
